requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-genai>=1.0.0
boto3>=1.34.0
ffmpeg-python>=0.2.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
faster-whisper>=1.0.0
mediapipe>=0.10.0
Pillow>=10.0.0
requests>=2.31.0
sqlalchemy>=2.0
alembic>=1.13
